requirements.txt
aiogram==3.0.0b8
requests
bs4
python-dotenv
fake-proxy
fake-useragent
selenium
selenium-wire